refactor & cleaning
