+++ imports.rb
@@ -8,8 +8,7 @@ require_relative 'assets/ascii'
 require_relative 'assets/cheat_mode'
 require_relative 'assets/names'
 require_relative 'assets/phrases'
-require_relative 'display/display_info'
-require_relative 'display/display_ui'
+require_relative 'display/display'
 require_relative 'display/menu'
 require_relative 'display/white_space'
 require_relative 'main/attack_mode'
